OMAP4+: EMIF: self-refresh as a single time device specific config
There is no need to reconfigure power management register multiple
times, instead choose a single value which will be configured for
the platform in use. This value is restored even after OFF mode,
in addition, we reduce the latency in the already critical configuration
path.

Since we may boot at different core OPP frequency, we depend on the
EMIF description to provide information about selfrefresh delay in
cycles directly.

We also allow for the emif device to now disable selfrefresh on need
by using -1.

Change-Id: I8b5025d7f28bd7f2b3ce5e35db71a5223973584d
Reported-by: Vinay Amancha <vinaykumar@ti.com>
Reported-by: Vinay Chaurasia <vinaych@ti.com>
Signed-off-by: Girish S G <girishsg@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>