tty: serial: omap: don't lock console in suspend.
serial_omap_suspend locks the console in suspend, even if it is
not being used as the console, or another console is present.  This
causes console messages to be lost if the device hangs in suspend.

The only reason to lock the console is to prevent new console writes
from turning the serial port hardware back on, which would prevent
the low power state in suspend.

To allow the console to continue functioning without leaving the
uart on in suspend, add a suspended flag to the port.  If the
uart is disabled while the suspend flag is set, use
pm_runtime_put_sync_suspend instead of pm_runtime_put_autosuspend
to force the uart to immediately go to a low power state.  This
will cause the uart to transition in and out of its low power
state for every console write after it suspends, but will allow
all console messages to immediately go to the serial console as
well as other active consoles.

Change-Id: Id8a380ceda56071efc8958acb27b5e7f76fe4b29
Signed-off-by: Colin Cross <ccross@android.com>