OMAP4: PM: Enable Auto VREF LPDDR I/O supply
We disabled Manual VREF CA and DQ control and depended on
boot loader configuration to setup configurations correctly.

Few boot loaders may miss this resulting in VREF supply CA and DQ
not being enabled. This may cause functional issue depending on
DDR behavior as if we disable Manual mode, auto mode is required
to be enabled.

Change-Id: I8033c101e45037a0201e23b8f462494e50f71bb8
Signed-off-by: Vinay Amancha <vinaykumar@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>