misc: fsa9480: Use external ID signal to verify USB attach
If the FSA9480 detection cycle starts while the ID signal is not
visible to the FSA9480 then the cable could be misidentified as
a USB-peripheral cable.  When this happens the FSA9480 detects
the VBUS but sees a floating ID signal.

To workaround this the FSA9480 driver verifies the dev_type using
the external ID signal when a USB peripheral cable has been
detected.  If there is a mismatch then the FSA9480 is reset to
start a new detection cycle.

Change-Id: Iefcf95780ebbf7c7f730264d48f88b532601f151
Signed-off-by: Adam Hampson <ahampson@sta.samsung.com>