ARM: omap4: cpuidle: prevent OSWR during idle
Prevent OSWR on mpuss during idle to allow the static dependency
between the mpuss and emif clock domains to be disabled.  A HW bug
prevents using OSWR if the static dependency is not enabled, but the
power savings of dropping the static dependency is higher than the
savings in OSWR vs. CSWR.

Change-Id: I7da24d2d29cb87c314bcf2db37ca668684383c7b
Signed-off-by: Colin Cross <ccross@android.com>