ARM: omap4: pm: remove static dependency between mpuss and emif
Remove the static dependency between mpuss and emif on 4460.
Requires OSWR in idle to be disabled.  A HW bug prevents using
OSWR if the static dependency is not enabled, but the power
savings of dropping the static dependency is higher than the
savings in OSWR vs. CSWR.

Based on patch by Ambresh K <ambresh@ti.com>

Change-Id: I97e32f560e1e889a6ea9434b1411b8d43c7f8604
Signed-off-by: Colin Cross <ccross@android.com>