ARM: omap: gpio: set wakeupenable bits when entering idle modes
THe pad wakeupenable bits are required to wake through a gpio when
in OFF or OSWR.  If the wakeupenable bits are set when the controller
is not in idle, a PRCM and a GPIO interrupt will occur for every
interrupt.  Set the wakeupenable bits for every active gpio
interrupt when entering any gpio idle mode.

Also sets the irq chip flag IRQCHIP_MASK_ON_SUSPEND to cause the
irq pm code to mask all non-wake gpios in suspend, which will
ensure the wakeupenable bit is not set on non-wake gpios.

Change-Id: I03b8d954a65f8cc47388319e983decf03308448f
Signed-off-by: Colin Cross <ccross@android.com>