i2c: omap3: Fix accessing of invalid registers
When cpu_is_omap44xx, dev->regs is set to point to
omap4_reg_map[], which contains entries for
OMAP_I2C_IRQENABLE_SET and OMAP_I2C_IRQENABLE_CLR.  These entries
are not present in the other reg_map.  However, the code that
uses these registers does not ensure that cpu_is_omap44xx before
using them.  This can cause register 0x00 (the revision register)
to be accessed instead and the proper register to not be
accessed.  This problem was observed on the OMAP3630 and caused
long hangs.

The fix is to ensure that cpu_is_omap44xx before accessing the
registers.

Thanks to vikram pandita for pointing me in the right direction.

Change-Id: Ic3accee9ce251724f13056bdb01f4db3308cd099
Signed-off-by: Scott Anderson <saa@google.com>