ASoC: McPDM: Bail out if module is not enabled
If McPDM MODULEMODE is in disable state, any interconnect access
to the module will result in an error and ultimately to a kernel
crash. If we hit that condition, it means that clocks are not
properly enabled and it is pointless to continue using the DAI,
so bailing out early, warning about this unexpected condition.

Change-Id: I476eddfdde9e9c82e3525b3723e9da8546c5c3da
Signed-off-by: PankajJindal <pankajjindal@ti.com>