OMAP3+: PM: DVFS: scale dependent voltage domains before the current domain
When scaling the device from Low OPP to High OPP, we currently do the
following sequence:
1- MPU voltage to OPP100+ voltage
2- CORE to opp 100
3- MPU frequency to OPP100+ frequency.
There is a small window between end of #1 and completion of #2, where
we have MPU @ OPP100 voltage CORE @ OPP50 voltage. Since, higher voltage
makes the transistors faster, even though we are at same OPP50 freq,
higher voltage will ensure that MPU will be faster and CORE will be
slower, This is where a timing violation can occur on the silicon.
Instead, it is safer to follow the sequence where each OPP
tuple is scaled together and there is no scope of timing violation:
- scale the dependent voltage domains voltage and frequency
- scale the current voltage domain voltage.
- scale the frequency.

Change-Id: I9966f6aae11c15cfb06420f668076aa56c2440a5
Signed-off-by: Avinash Mahadeva <avinashhm@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>