ARM: omap4: PM: mask all PPIs when going into low power mode
This is primarily needed to mask the local timer irq that could be pending
since timekeeping gets suspended after the local irqs are disabled.
The pending interrupt would kick the CPU out of WFI immediately, and prevent
it from getting to the lower power state. The correct value will be restored
when the CPU is brought back up by restore.

Change-Id: I6620bff453131257923bcc0e322399e9b99e28c3
Signed-off-by: Dima Zavin <dima@android.com>