ASoC: twl6040: Prevent extra power-up/down during resume
TWL6040 CODEC will reach BIAS_OFF state while suspended, it can be
directly in suspend handler or if the CODEC was idle at the time
suspend came. Currently in the latter case, there is an unnecessary
transition to BIAS_STANDBY and then back to BIAS_OFF while resuming.
BIAS_STANDBY puts TWL6040 hardware into Power-On state, so it's a
costly operation which is not required in the case described above.

Change-Id: Ic635e191f6dbd3904a9994a66e54bcb615ee2cdb
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>