gpio: omap: fix gpio transitions to off mode
When transitioning to off mode, it is possible that the powerdomain
that contains the gpio block does not transition due to another
clock domain being on.  In this case, we must do the same
configuration on the gpios as in the non-off mode case.

Modify omap2_gpio_set_edge_wakeup to use different variables
to save context vs. omap_save_context, and both set the edge
wakeups and call pm_runtime_put_sync when going to off mode.

Change-Id: I2890f7f3dd2c3dfe493cf748a523004a76e560c2
Signed-off-by: Colin Cross <ccross@android.com>