OMAP4: PM: program warm_reset latency
Warm reset could be triggered by normal reboot sequence or with
various OMAP watchdog timers. Watchdog timers introduce unplanned
reset sequence which could, in theory occur at any point of time.

OMAP provides two registers to control the length that reset is
held low allowing for the voltages and oscillator to reach stable
state. PRM_RSTTIME::RSTTIME1 which is a default duration which is
added and PRM_VOLTSETUP_WARMRESET depending on which voltage domains
are in LP state. Considering watchdog, it is really not practical
to optimally program PRM_VOLTSETUP_WARMRESET, hence we use the
worst case latency in the worst case scenario in RSTTIME1.

From a latency consideration, system restarting from an OFF mode
would be the worst case for watchdog. Here, we need to consider
ramping of each voltage domains + Oscillator which could have been
caught at the worst possible time - at the start of oscillator
shutdown sequence.

Since all these information is already available in the
datastructures maintained by PM framework, it is not necessary for
board files to provide any further information.

Change-Id: I2d2213d44b7d4381bdc2acf4ad81b6dc432d8f6b
Reported-by: Baek.Kyung-Han <wildtaz.baek@samsung.com>
Signed-off-by: Nishanth Menon <nm@ti.com>