(TEMP) omap: rpmsg: add mutex to protect mbox handle
In order to protect the omap_mbox_send function againts a omap_mbox_put
which is done on suspend callback, we are adding a mutex to avoid a NULL
pointer dereference.

For some reason (still under investigation) some time the mailbox handles are
null even after calling rproc_last_busy.  All the code seems to cover all
possible race conditions, but looks like there's something wrong in the
pm_runtime function with calculating the expiration time because it is not
catching the jiffies wrap. It is difficult to reproduce, meanwhile this patch
is made to recover from that situation calling rproc_last_busy again. In other
to wake up ducati and therefore the mailbox.

Once the proper fix is checked into linux-omap-3.0, we will revert this
change.

Change-Id: I609ac1c2baabbbb8886e621a95a05f5726d329d2
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Iliyan Malchev <malchev@google.com>