OMAP4+: HSI: fix for spurious CAWAKE event while processing RX/TX
When a CAWAKE event occurs while processing a RX or TX interrupt, the backup
interrupt status is set along with normal interrupt status bit.
So HSI driver detects this as a new CAWAKE interrupt and process it.
But because normal HSI interrupt has also been set, the tasklet is rescheduled
and the same CAWAKE event is processed again, causing spurious CAWAKE
notifications.

Fix consists in adding a check to process a new CAWAKE event as a double
interrupt only if normal CAWAKE interrupt is currently being processed.
Otherwise, let the taslket be rescheduled to process new interrupt.

Bug: 5956850, 5975406
Change-Id: Ia474ea1bc30da12469cad1f4aea2547b1cf9d244
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com
Signed-off-by: wonhee <wonhee48.seo@samsung.com>