pm44xx: Fix iva and dsp modules being stuck at high frequency
I discovered a hideous bug where the IVA and DSP omap4 devices are
stuck at highest frequency at boot until there's a resquestor.

So this isn't a huge problem with IVA since a lot of apps are valid
requestors, but there doesn't really seem to be a valid requestor
for DSP, so it's pretty much stuck at highest frequency while it's
awake.

This patch fixes it.