pm44xx: revert force setting of DSP freq at boot but leave IVA
This DSP device is a bit of a mystery. The device gets created, but
completely ignores OPP table, never sees any requestors, and basically
never gets used. I'm going to leave it alone for now - I fear force
setting the frequency could have a detrimental effect.  I'll take a
closer look at it later.

Left the IVA fix in there because it's an awesome fix.